+++ slop/crates/veil/src/zk/verifier_ctx.rs
@@ -132,9 +132,9 @@ impl<GC: ZkIopCtx> ReadingCtx for ZkVerifierCtx<GC> {
         Ok(())
     }
 
-    fn read_oracle(&mut self, log_width: usize, log_stacking: usize) -> Option<MleCommit> {
+    fn read_oracle(&mut self, log_width: u32, log_stacking: u32) -> Option<MleCommit> {
         self.inner
-            .read_next_pcs_commitment(log_width, log_stacking)
+            .read_next_pcs_commitment(log_width as usize, log_stacking as usize)
             .map(|idx| MleCommit { inner: idx })
     }
